requirements/common.txt
# Common requirements for all environments

django == 4.1.2
django-redis == 5.2.0
djangorestframework == 3.14.0
drf-nested-routers == 0.93.4
psycopg2-binary == 2.9.5